requirements-dev.txt
# Testing dependencies
pytest==7.4.0
pytest-flask==1.2.0
pytest-mock==3.11.1
coverage==7.3.0